% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_koppen_sthao.R
\name{classify_koppen}
\alias{classify_koppen}
\title{Koppen-Geiger climate classification}
\usage{
classify_koppen(Tmonth, Pmonth, lat)
}
\arguments{
\item{Tmonth}{A numerical vector of length 12. 
It contains near-surface temperature value for each month of the year
(January to December). Temperature is expressed in degrees Celsius.}

\item{Pmonth}{A numerical vector of length 12. 
It contains the value of precipitation for each month of the year
(January to December). Precipitation is expressed in millimeters.}

\item{lat}{the latitude of location for which \code{Tmonth} and \code{Pmonth} are provided.}
}
\value{
A string with the name of the Koppen-Geiger class.
}
\description{
Find the Koppen-Geiger class for one location.
}
